+++ front/lib/features/reports/presentation/reports_page.dart
@@ -1,6 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:front/features/reports/presentation/reports_form.dart';
-
 import '../../../common/background/background.dart';
 
 class ReportsPage extends StatelessWidget {
@@ -12,7 +10,6 @@ class ReportsPage extends StatelessWidget {
       body: Stack(
         children: [
           GradientBackground(),
-          ReportsForm(),
         ],
       ),
     );
